split off config options from process lib
