+++ src/lib/qr_alchemy/gui_config.py
@@ -6,6 +6,7 @@ from gi.repository import Gtk,Gio
 import qr_alchemy.process as qr_process
 import qr_alchemy.plugins as qr_plugins
 import qr_alchemy.gui as gui
+import qr_alchemy.config as qr_config
 
 
 class QRConfig(Gtk.Window):
@@ -117,7 +118,7 @@ class QRConfig(Gtk.Window):
         if state != Gtk.ResponseType.OK:
             return
 
-        qr_process.qr_update_configaction(data[i][0], '', '')
+        qr_config.qr_update_configaction(data[i][0], '', '')
         self.ls_actions_populate(self.ls_act)
         
     def bu_add_clicked(self, qr_code):
@@ -364,7 +365,7 @@ class QRConfigEntry(Gtk.Dialog):
             subtype=self.prog
         else:
             subtype=''
-        qr_process.qr_update_configaction(self.code_type, self.action_type, subtype)
+        qr_config.qr_update_configaction(self.code_type, self.action_type, subtype)
         self.destroy()
         
     def bu_cancel_clicked(self, button):
@@ -389,7 +390,7 @@ class QRConfigEntry(Gtk.Dialog):
         if state != Gtk.ResponseType.OK:
             return
 
-        qr_process.qr_update_configaction(self.code_type, '', '')
+        qr_config.qr_update_configaction(self.code_type, '', '')
         self.state = Gtk.ResponseType.OK
         self.destroy()
     def get_state():
